feat(ui): preview Kubernetes contexts after kubeconfig upload
