+++ ui/components/connections/wizard/kubernetesExtension.tsx
@@ -1,49 +1,25 @@
 import { useMemo, useState } from 'react';
 import Link from 'next/link';
 import {
-  Alert,
   Box,
-  Checkbox,
   CheckCircleIcon,
   Chip,
   CircularProgress,
-  CustomTooltip,
-  MenuItem,
-  TextField,
   Typography,
-  CloudUploadIcon,
-  AssignmentTurnedInIcon,
   SettingsIcon,
 } from '@sistent/sistent';
 import { useSelector } from 'react-redux';
 import { alpha, styled } from '@/theme';
-import { EVENT_TYPES } from 'lib/event-types';
 import { CONNECTION_STATES } from '@/utils/Enum';
-import {
-  CONNECTIONS_PATH,
-  getKubernetesContexts,
-  kubernetesImportedNotify,
-} from '../ConnectionWizard.helpers';
-import { formatWizardError } from './errors';
-import {
-  DEFAULT_MESHSYNC_DEPLOYMENT_MODE,
-  getMeshsyncModeTooltip,
-  MESHSYNC_DEPLOYMENT_MODE_OPTIONS,
-  MESHSYNC_MODES_DOCS_URL,
-} from './kubernetesDeploymentMode';
+import { CONNECTIONS_PATH, getKubernetesContexts } from '../ConnectionWizard.helpers';
 import { kubernetesSettingsStep } from './kubernetesSettings';
 import FormatConnectionMetadata from '../metadata';
 import { ConnectionStateChip } from '../ConnectionChip';
-import { KubernetesImportStep, StepHeader } from '../ConnectionWizardStepContent';
-import type {
-  ConnectionExtension,
-  DiscoveredKubeContext,
-  GenericRecord,
-  WizardContext,
-  WizardStep,
-} from './types';
+import { StepHeader } from '../ConnectionWizardStepContent';
+import { kubernetesDetailsStep } from './kubernetesImportStep';
+import { kubernetesReviewStep } from './kubernetesReviewStep';
+import type { ConnectionExtension, GenericRecord, WizardContext, WizardStep } from './types';
 
-const KUBECONFIG_DOCS_URL = 'https://docs.meshery.io/installation/kubernetes';
 const KUBERNETES_CONNECTION_DOCS_URL =
   'https://docs.meshery.io/guides/infrastructure-management/kubernetes-connection-lifecycle';
 
@@ -62,16 +38,6 @@ const ContextRow = styled(Box, {
   '&:hover': {
     borderColor: theme.palette.background.brand?.default,
   },
-}));
-
-const ConnectNotice = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  gap: theme.spacing(1),
-  borderRadius: theme.spacing(1),
-  border: `1px solid ${theme.palette.divider}`,
-  background: alpha(theme.palette.info.main, 0.06),
-  padding: theme.spacing(1, 1.5),
 }));
 
 const SuccessBadge = styled(Box)(({ theme }) => ({
@@ -96,327 +62,6 @@ const ConnectionsLink = styled(Link)(({ theme }) => ({
   textDecoration: 'underline',
   textUnderlineOffset: 2,
 }));
-
-type ContextChoice = { selected: boolean; name: string; meshsyncDeploymentMode: string };
-
-// ---------------------------------------------------------------------------
-// postConfig accessors. The kubernetes flow keeps its working state on the
-// wizard's free-form postConfig bag.
-// ---------------------------------------------------------------------------
-
-const getDiscovered = (ctx: WizardContext): DiscoveredKubeContext[] =>
-  (ctx.data.postConfig.discoveredContexts as DiscoveredKubeContext[]) || [];
-
-const getChoices = (ctx: WizardContext): Record<string, ContextChoice> =>
-  (ctx.data.postConfig.contextChoices as Record<string, ContextChoice>) || {};
-
-const getConnectOnImport = (ctx: WizardContext): boolean =>
-  ctx.data.postConfig.connectOnImport !== false;
-
-// A discovered context's choice, defaulted for contexts not yet touched.
-const defaultChoice = (name: string): ContextChoice => ({
-  selected: true,
-  name,
-  meshsyncDeploymentMode: DEFAULT_MESHSYNC_DEPLOYMENT_MODE,
-});
-
-// ---------------------------------------------------------------------------
-// 2. Import kubeconfig (overrides the generic details step). Picks a file and,
-// on Next, discovers its contexts WITHOUT persisting them.
-// ---------------------------------------------------------------------------
-
-const KubeconfigStepBody = ({ ctx }: { ctx: WizardContext }) => (
-  <KubernetesImportStep
-    kubeconfigFile={ctx.data.kubeconfigFile}
-    onPickFile={(kubeconfigFile) => {
-      ctx.patch({ kubeconfigFile, registrationResult: null, registrationError: null });
-      // Drop any previously discovered contexts so a new file is re-discovered.
-      ctx.patchPostConfig({ discoveredContexts: undefined, contextChoices: undefined });
-    }}
-  />
-);
-
-const kubernetesDetailsStep: WizardStep = {
-  id: 'kubeconfig',
-  label: 'Import Kubeconfig',
-  icon: CloudUploadIcon,
-  Component: KubeconfigStepBody,
-  canProceed: (ctx) => Boolean(ctx.data.kubeconfigFile),
-  nextLabel: () => 'Continue',
-  helpText: `Upload a kubeconfig with embedded certificates. Meshery reads the file's contexts, then registers the ones you select as Kubernetes connections. [Learn more about connecting Kubernetes](${KUBECONFIG_DOCS_URL}).`,
-  onNext: async (ctx) => {
-    if (!ctx.data.kubeconfigFile) {
-      return false;
-    }
-    ctx.patch({ registrationError: null });
-    try {
-      const discovered = await ctx.services.discoverKubeContexts(ctx.data.kubeconfigFile);
-      if (discovered.length === 0) {
-        ctx.services.notify({
-          message: 'No Kubernetes contexts were found in the kubeconfig.',
-          event_type: EVENT_TYPES.WARNING,
-        });
-        return false;
-      }
-      const contextChoices = Object.fromEntries(
-        discovered.map((context) => [context.id, defaultChoice(context.name)]),
-      );
-      ctx.patchPostConfig({
-        discoveredContexts: discovered,
-        contextChoices,
-        connectOnImport: true,
-      });
-      return true;
-    } catch (error) {
-      ctx.patch({ registrationError: error });
-      ctx.services.notify({
-        message: `Failed to read kubeconfig: ${formatWizardError(error)}`,
-        event_type: EVENT_TYPES.ERROR,
-      });
-      return false;
-    }
-  },
-};
-
-// ---------------------------------------------------------------------------
-// 3. Review contexts (overrides the generic register step). Select which
-// discovered contexts to import, rename them, and choose whether reachable
-// clusters are connected on import. On Next the selected contexts are created
-// and the reachable ones (when chosen) are transitioned to connected.
-// ---------------------------------------------------------------------------
-
-const ReviewContextsStepBody = ({ ctx }: { ctx: WizardContext }) => {
-  const discovered = getDiscovered(ctx);
-  const choices = getChoices(ctx);
-  const connectOnImport = getConnectOnImport(ctx);
-  const hasReachable = discovered.some((context) => context.reachable);
-
-  const updateChoice = (id: string, partial: Partial<ContextChoice>) =>
-    ctx.patchPostConfig({
-      contextChoices: { ...choices, [id]: { ...choices[id], ...partial } },
-    });
-
-  return (
-    <Box sx={{ display: 'grid', gap: 3 }}>
-      <StepHeader
-        title="Review contexts"
-        subtitle="Choose which clusters to import, rename them if you like, and decide whether reachable clusters should be connected."
-      />
-      {Boolean(ctx.data.registrationError) && (
-        <Alert severity="error" variant="filled">
-          {formatWizardError(ctx.data.registrationError)}
-        </Alert>
-      )}
-      <Box sx={{ display: 'grid', gap: 1.5 }}>
-        {discovered.map((context) => {
-          const choice = choices[context.id] || defaultChoice(context.name);
-          return (
-            <ContextRow key={context.id} muted={!choice.selected}>
-              <Checkbox
-                checked={choice.selected}
-                onChange={(event) => updateChoice(context.id, { selected: event.target.checked })}
-                sx={{ p: 0 }}
-              />
-              {/* data-testid anchors each discovered context row for e2e. */}
-              <Box sx={{ flex: 1, minWidth: 0 }} data-testid="connection-wizard-context-row">
-                <TextField
-                  variant="standard"
-                  fullWidth
-                  value={choice.name}
-                  disabled={!choice.selected}
-                  onChange={(event) => updateChoice(context.id, { name: event.target.value })}
-                  inputProps={{
-                    'aria-label': `Name for ${context.name}`,
-                    title: choice.name,
-                  }}
-                  sx={{
-                    maxWidth: '100%',
-                    '& .MuiInput-input': {
-                      overflow: 'hidden',
-                      textOverflow: 'ellipsis',
-                      whiteSpace: 'nowrap',
-                    },
-                    '& .MuiInput-underline:before, & .MuiInput-underline:after': {
-                      maxWidth: '100%',
-                    },
-                  }}
-                />
-                <Typography variant="caption" color="text.secondary" noWrap component="div">
-                  {context.server || 'unknown server'}
-                </Typography>
-              </Box>
-              {/* MeshSync mode per context; hover shows details + docs link. */}
-              <TextField
-                select
-                variant="standard"
-                label="MeshSync"
-                value={choice.meshsyncDeploymentMode}
-                disabled={!choice.selected}
-                onChange={(event) =>
-                  updateChoice(context.id, { meshsyncDeploymentMode: event.target.value })
-                }
-                sx={{ minWidth: 120 }}
-                inputProps={{ 'aria-label': `MeshSync deployment mode for ${context.name}` }}
-              >
-                {MESHSYNC_DEPLOYMENT_MODE_OPTIONS.map((option) => (
-                  <MenuItem key={option.value} value={option.value}>
-                    <CustomTooltip
-                      interactive
-                      title={getMeshsyncModeTooltip(option)}
-                      placement="left"
-                    >
-                      <Box component="span" sx={{ display: 'block', width: '100%' }}>
-                        {option.label}
-                      </Box>
-                    </CustomTooltip>
-                  </MenuItem>
-                ))}
-              </TextField>
-              <ConnectionStateChip
-                status={
-                  context.reachable ? CONNECTION_STATES.DISCOVERED : CONNECTION_STATES.NOTFOUND
-                }
-              />
-            </ContextRow>
-          );
-        })}
-      </Box>
-      <ConnectNotice
-        onClick={() => hasReachable && ctx.patchPostConfig({ connectOnImport: !connectOnImport })}
-        sx={{ cursor: hasReachable ? 'pointer' : 'default' }}
-      >
-        <Checkbox
-          checked={connectOnImport}
-          disabled={!hasReachable}
-          onChange={(event) => ctx.patchPostConfig({ connectOnImport: event.target.checked })}
-          onClick={(event) => event.stopPropagation()}
-          sx={{ p: 0 }}
-        />
-        <Typography variant="body2">
-          Connect reachable clusters after import. Unreachable clusters are imported as discovered
-          and can be connected later.
-        </Typography>
-      </ConnectNotice>
-    </Box>
-  );
-};
-
-/** Connections returned by the upload endpoint, with how they should be acted on. */
-type CreatedContext = {
-  connectionId: string;
-  name: string;
-  reachable: boolean;
-  alreadyConnected: boolean;
-};
-
-const collectCreated = (result: GenericRecord | null): CreatedContext[] => {
-  const toCreated = (records: GenericRecord[], alreadyConnected: boolean): CreatedContext[] =>
-    records.map((record) => ({
-      connectionId: String(record.connectionId ?? record.id ?? ''),
-      name: String(record.name ?? ''),
-      reachable: Boolean(record.reachable),
-      alreadyConnected,
-    }));
-
-  return [
-    ...toCreated(getKubernetesContexts(result, 'registered'), false),
-    ...toCreated(getKubernetesContexts(result, 'connected'), true),
-  ];
-};
-
-const kubernetesReviewStep: WizardStep = {
-  id: 'kubernetes-review',
-  label: 'Review Contexts',
-  icon: AssignmentTurnedInIcon,
-  Component: ReviewContextsStepBody,
-  nextLabel: () => 'Import',
-  helpText: `Select which contexts to import, choose a MeshSync mode (embedded or operator), and decide whether reachable clusters should connect immediately. Hover a MeshSync option for details. [MeshSync modes](${MESHSYNC_MODES_DOCS_URL}). [Kubernetes connection lifecycle](${KUBERNETES_CONNECTION_DOCS_URL}).`,
-  canProceed: (ctx) => getDiscovered(ctx).some((context) => getChoices(ctx)[context.id]?.selected),
-  onNext: async (ctx) => {
-    const discovered = getDiscovered(ctx);
-    const choices = getChoices(ctx);
-    const selected = discovered.filter((context) => choices[context.id]?.selected);
-    if (!ctx.data.kubeconfigFile || selected.length === 0) {
-      return false;
-    }
-    ctx.patch({ registrationError: null });
-    try {
-      const contexts = Object.fromEntries(
-        selected.map((context) => {
-          const choice = choices[context.id];
-          return [
-            context.id,
-            {
-              name: choice.name.trim() || context.name,
-              meshsyncDeploymentMode: choice.meshsyncDeploymentMode,
-            },
-          ];
-        }),
-      );
-      const result = await ctx.services.uploadKubeconfig(ctx.data.kubeconfigFile, {
-        selectedContextIds: selected.map((context) => context.id),
-        contexts,
-      });
-      ctx.patch({ registrationResult: result ?? {} });
-
-      const created = collectCreated(result ?? {});
-      const connectedIds = new Set(
-        created
-          .filter((context) => context.alreadyConnected)
-          .map((context) => context.connectionId),
-      );
-      if (getConnectOnImport(ctx)) {
-        const toConnect = created.filter(
-          (context) => context.reachable && !context.alreadyConnected && context.connectionId,
-        );
-        const outcomes = await Promise.allSettled(
-          toConnect.map((context) =>
-            ctx.services.updateConnectionById(context.connectionId, {
-              status: CONNECTION_STATES.CONNECTED,
-            }),
-          ),
-        );
-        toConnect.forEach((context, index) => {
-          if (outcomes[index].status === 'fulfilled') {
-            connectedIds.add(context.connectionId);
-          }
-        });
-      }
-
-      // Per-context final state, mapped to the canonical connection states so
-      // the receipt can reuse the shared status chip.
-      const importedContexts = created.map((context) => ({
-        name: context.name,
-        status: connectedIds.has(context.connectionId)
-          ? CONNECTION_STATES.CONNECTED
-          : context.reachable
-            ? CONNECTION_STATES.DISCOVERED
-            : CONNECTION_STATES.NOTFOUND,
-      }));
-
-      ctx.patchPostConfig({
-        importedContexts,
-        createdCount: created.length,
-        connectedCount: connectedIds.size,
-        unreachableCount: created.filter((context) => !context.reachable).length,
-      });
-      ctx.services.notify(kubernetesImportedNotify(created.length));
-      return true;
-    } catch (error) {
-      ctx.patch({ registrationError: error });
-      ctx.services.notify({
-        message: `Failed to import kubeconfig: ${formatWizardError(error)}`,
-        event_type: EVENT_TYPES.ERROR,
-      });
-      return false;
-    }
-  },
-};
-
-// ---------------------------------------------------------------------------
-// Post-config (configure mode only): act on an already-registered connection's
-// contexts. Hidden during creation, where the review step covers connecting.
-// ---------------------------------------------------------------------------
 
 type DiscoveredContext = {
   connectionId: string;
@@ -454,7 +99,6 @@ const ContextsStepBody = ({ ctx }: { ctx: WizardContext }) => {
     () => collectContexts(ctx.data.registrationResult),
     [ctx.data.registrationResult],
   );
-  // Per-context status overrides applied during this session.
   const overrides = (ctx.data.postConfig.contextStatuses as Record<string, string>) || {};
   const [busyId, setBusyId] = useState<string | null>(null);
 
@@ -470,8 +114,8 @@ const ContextsStepBody = ({ ctx }: { ctx: WizardContext }) => {
       });
     } catch (error) {
       ctx.services.notify({
-        message: `Failed to update ${context.name}: ${formatWizardError(error)}`,
-        event_type: EVENT_TYPES.ERROR,
+        message: `Failed to update ${context.name}: ${error}`,
+        event_type: 'error',
       });
     } finally {
       setBusyId(null);
@@ -553,25 +197,15 @@ const kubernetesContextsStep: WizardStep = {
   label: 'Manage Clusters',
   icon: SettingsIcon,
   Component: ContextsStepBody,
-  // The creation flow handles selecting + connecting in the review step; this
-  // step only adds value when (re)configuring an existing connection that
-  // actually carries discovered contexts to manage.
   hidden: (ctx) =>
     ctx.mode === 'create' || collectContexts(ctx.data.registrationResult).length === 0,
 };
-
-// ---------------------------------------------------------------------------
-// 5. Receipt
-// ---------------------------------------------------------------------------
 
 const KubernetesReceiptBody = ({ ctx }: { ctx: WizardContext }) => {
   const controllerState = useSelector(
     (state: { ui: { controllerState: unknown } }) => state.ui.controllerState,
   );
 
-  // Configure mode operates on a single connection: show the same live detail
-  // "receipt" that the Connections table renders when a row is expanded —
-  // status chips, controller versions, and the Diagnostics section.
   const configuredConnection = (ctx.data.registrationResult as GenericRecord) || {};
   if (
     ctx.mode === 'configure' &&
@@ -668,7 +302,7 @@ const kubernetesReceiptStep: WizardStep = {
 export const kubernetesExtension: ConnectionExtension = {
   match: { kind: 'kubernetes' },
   detailsStep: kubernetesDetailsStep,
-  credentialStep: null, // the kubeconfig is the credential
+  credentialStep: null,
   registerStep: kubernetesReviewStep,
   postConfigSteps: [kubernetesSettingsStep, kubernetesContextsStep],
   receiptStep: kubernetesReceiptStep,
